Move FirebaseFunctionRequest to shared data module
