--- shared/features/game/src/commonMain/kotlin/com/yral/shared/features/game/data/models/FirebaseFunctionRequest.kt
@@ -1,10 +1,0 @@
-package com.yral.shared.features.game.data.models
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-internal data class FirebaseFunctionRequest<T>(
-    @SerialName("data")
-    val data: T,
-)
